requirements.txt
art==5.6
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
